refactor(uc-daemon): extract GUI-agnostic daemon runtime into uc-daemon crate (ADR-008 P1)
把 uc-desktop/src/daemon 的 GUI-agnostic runtime（27 模块：run_mode/service/handle/ownership/state/tokio_runtime + workers/peers/search + bootstrap/run_loop/startup_recovery/各装配 + app/app_assembly/mobile_lan_lifecycle）迁入新建 uc-daemon crate。uc-desktop 留 host 胶水（host.rs：start_in_process/ProcessRuntimeHandles/run* 原地拆分保留，内部前向 use uc_daemon::daemon::*），daemon/mod.rs re-export 保 uc_desktop::daemon::* 接口面——uc-cli/uc-tauri 零改动。DaemonRunMode::GuiInProcess 迁入但不删（D2 删除留 P3）。uc-daemon 零反依赖 uc-desktop。纯结构迁移、零行为变化：cargo check --workspace 通过、48 迁移单测全绿、daemon_probe 测试编译通过。
