+++ src-tauri/crates/uc-daemon/src/daemon/app_assembly.rs
