+++ src-tauri/crates/uc-daemon/src/daemon/app_facade_assembly.rs
