+++ src-tauri/crates/uc-daemon/src/daemon/search/coordinator.rs
