+++ src-tauri/crates/uc-daemon/src/daemon/search_assembly.rs
