+++ src-tauri/crates/uc-daemon/src/daemon/service.rs
