+++ src-tauri/crates/uc-daemon/src/daemon/service_assembly.rs
