+++ src-tauri/crates/uc-daemon/src/daemon/startup_recovery.rs
