+++ src-tauri/crates/uc-daemon/src/daemon/state.rs
